src/requirements.txt
requests
psycopg2-binary
pandas
python-dateutil
watchdog